Dashboard e historico com component de mes/ano

+++ Components/Dashboard/Model/DashboardResponse.cs
@@ -4,5 +4,5 @@ namespace FinancialManager.Web.Components.Dashboard.Model
         decimal Amount,
         decimal Entrys,
         decimal Outputs,
-        IList<Tuple<string, double>> Categories);
+        IList<Tuple<string, decimal>> Categories);
 }
